Move games-specific models into subpackages
Closes: https://github.com/nc0fr/conception-application/issues/18
