+++ src/main/java/fr/nc0/cda/modele/nim/EtatPartieNim.java
@@ -4,7 +4,7 @@
  * that can be found in the COPYRIGHT file.
  */
 
-package fr.nc0.cda.modele;
+package fr.nc0.cda.modele.nim;
 
 /** Représente l'état d'une partie du jeu de Nim. */
 public enum EtatPartieNim {
