+++ src/main/java/fr/nc0/cda/modele/nim/Nim.java
@@ -4,7 +4,7 @@
  * that can be found in the COPYRIGHT file.
  */
 
-package fr.nc0.cda.modele;
+package fr.nc0.cda.modele.nim;
 
 import java.util.ArrayList;
 import java.util.Iterator;
