+++ src/main/java/fr/nc0/cda/modele/puissance4/CellulePuissance4.java
@@ -4,7 +4,7 @@
  * that can be found in the COPYRIGHT file.
  */
 
-package fr.nc0.cda.modele;
+package fr.nc0.cda.modele.puissance4;
 
 /** Enumération des différentes cellules possibles dans une grille de Puissance 4. */
 public enum CellulePuissance4 {
